fix: [feature/video] - VideoList 컴포넌트에서 카테고리 필드 분리 등 코드 정리
