+++ web/src/components/ExerciseInfo/VideoPopup.tsx
@@ -1,5 +1,29 @@
 import styled from '@emotion/styled';
 
+type VideoPopupProps = {
+    video: {
+        id: string;
+        title: string;
+    };
+    onClose: () => void;
+};
+function VideoPopup({ video, onClose }: VideoPopupProps) {
+    return (
+        <VideoPopupContainer onClick={onClose}>
+            {/* <CloseButton onClick={onClose}>닫기</CloseButton> */}
+            <iframe
+                width="560"
+                height="315"
+                src={`https://www.youtube.com/embed/${video.id}`}
+                title={video.title}
+                frameBorder="0"
+                allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+                allowFullScreen
+            ></iframe>
+        </VideoPopupContainer>
+    );
+}
+
 const VideoPopupContainer = styled.div`
     position: fixed;
     top: 0;
@@ -18,28 +42,5 @@ const VideoPopupContainer = styled.div`
 //     top: 20px;
 //     right: 20px;
 // `;
-type VideoPopupProps = {
-    video: {
-        id: string;
-        title: string;
-    };
-    onClose: () => void;
-};
-function VideoPopup({ video, onClose }: VideoPopupProps) {
-    return (
-        <VideoPopupContainer onClick={onClose}>
-            {/* <CloseButton onClick={onClose}>닫기</CloseButton> */}
-            <iframe 
-                width="560" 
-                height="315" 
-                src={`https://www.youtube.com/embed/${video.id}`} 
-                title={video.title} 
-                frameBorder="0" 
-                allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" 
-                allowFullScreen>
-            </iframe>
-        </VideoPopupContainer>
-    );
-}
 
 export default VideoPopup;
