+++ web/src/components/ExerciseInfo/YoutubeApi.tsx
@@ -28,8 +28,8 @@ export type VideosResponse = {
 // export const YOUTUBE_API_KEY = 'AIzaSyDPiKdEMBJDeWYQwijKyYRw_VVO0ePIr1c';
 // export const YOUTUBE_API_KEY = 'AIzaSyD26946NOfQsR5aSkR4cfZnx5sUq-B-dTg';
 // export const YOUTUBE_API_KEY = 'AIzaSyASbmi09zmEJHh-6wd9j9F40j2519wYGoo';
-export const YOUTUBE_API_KEY = 'AIzaSyAIXJnQWRynbV5Lg1f1lRGTz95vYzj8WUM';
-// export const YOUTUBE_API_KEY = 'AIzaSyB6Es_UBvC3ygEEGaruLVXDCX6OyUWIKPE';
+// export const YOUTUBE_API_KEY = 'AIzaSyAIXJnQWRynbV5Lg1f1lRGTz95vYzj8WUM';
+export const YOUTUBE_API_KEY = 'AIzaSyB6Es_UBvC3ygEEGaruLVXDCX6OyUWIKPE';
 // export const YOUTUBE_API_KEY = 'AIzaSyAGwV_z2kIHmQKYqgpodo1QyFpt2O2mADk';
 // export const YOUTUBE_API_KEY = 'AIzaSyDoSnXhlTJyL0bXj3p52a1q7HsOkGS2Y4s';
 // export const YOUTUBE_API_KEY = 'AIzaSyCiNRWlBZEQLTiZczKSieY7BS8lOLAA3_8';
