+++ web/src/pages/ExerciseInfo.tsx
@@ -1,11 +1,41 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import styled from '@emotion/styled';
+import { useRecoilState } from 'recoil';
+import { categoryRecoil } from '../recoil/video/atoms';
 import VideoList from '../components/ExerciseInfo/VideoList';
+import { resetTotalResults } from '../components/ExerciseInfo/YoutubeApi';
 
 const ExerciseInfo: React.FC = () => {
+    const [category, setCategory] = useRecoilState<string>(categoryRecoil);
+
+    const handleTabClick = useCallback((newCategory: string) => {
+        setCategory('');
+        resetTotalResults();
+        setCategory(newCategory);
+    }, []);
     return (
         <ExerciseInfoPage>
             <PageTitle>운동 정보</PageTitle>
+            <BtnTab>
+                <button
+                    className={`category01 ${category === 'running' ? 'active' : ''}`}
+                    onClick={() => handleTabClick('running')}
+                >
+                    러닝
+                </button>
+                <button
+                    className={`category02 ${category === 'hiking' ? 'active' : ''}`}
+                    onClick={() => handleTabClick('hiking')}
+                >
+                    등산
+                </button>
+                <button
+                    className={`category03 ${category === 'health' ? 'active' : ''}`}
+                    onClick={() => handleTabClick('health')}
+                >
+                    헬스
+                </button>
+            </BtnTab>
             <VideoList />
         </ExerciseInfoPage>
     );
@@ -30,6 +60,37 @@ const PageTitle = styled.h2`
         height: 1px;
         color: #000;
         background-color: #000;
+    }
+`;
+
+const BtnTab = styled.div`
+    margin-top: 10px;
+    position: relative;
+    top: 60px;
+    z-index: 10;
+
+    button {
+        position: absolute;
+        left: 50%;
+        transform: translateX(-50%);
+        border-style: none;
+        border-radius: 15px;
+        padding: 3px 10px;
+        background-color: #fff;
+        box-shadow: 2.5px 5px 10px rgba(0, 0, 0, 0.5);
+
+        &.active {
+            background-color: #000;
+            color: #fff;
+        }
+    }
+    .category01 {
+        left: 43.75%;
+        transform: translateX(-40%);
+    }
+    .category03 {
+        left: 56.3%;
+        transform: translateX(-60%);
     }
 `;
