+++ web/src/recoil/video/atoms.ts
@@ -1,17 +1,16 @@
 import { atom } from 'recoil';
-import { Video } from '../../components/ExerciseInfo/YoutubeApi';
 
 interface Playlist {
     playlistName: string;
     userId: number;
 }
 
-export const bookmarkedVideosState = atom<Video[]>({
-    key: 'bookmarkedVideosState',
-    default: [],
-});
-
 export const playlistsDataRecoil = atom<Playlist[] | null>({
     key: 'playlistsDataRecoil',
     default: null,
+});
+
+export const categoryRecoil = atom<string>({
+    key: 'categoryRecoil',
+    default: 'running',
 });
